requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0
jupyter>=1.0.0
python-dotenv>=0.19.0
mlflow>=1.23.0
joblib>=1.1.0
